[BE]merge

See merge request s11-webmobile1-sub2/S11P12D104!78

+++ backend/src/main/java/com/tunapearl/saturi/domain/lesson/LessonGroupResultEntity.java
@@ -24,9 +24,9 @@ public class LessonGroupResultEntity {
     @JoinColumn(name = "lesson_group_id")
     private LessonGroupEntity lessonGroup;
 
-    private Integer avgSimilarity; // 평균 유사도
+    private Long avgSimilarity; // 평균 유사도
 
-    private Integer avgAccuracy; // 평균 정확도
+    private Long avgAccuracy; // 평균 정확도
 
     private LocalDateTime startDt;
